Resolve possible issues when comparing signed and unsigned integer

+++ src/botlib/be_ai_char.cpp
@@ -230,7 +230,7 @@ bot_character_t *BotLoadCharacterFromFile(char *charfile, int skill)
 				return NULL;
 			} //end if
 			//if it's the correct skill
-			if (skill < 0 || token.intvalue == skill)
+			if (skill < 0 || token.intvalue == (unsigned long)skill)
 			{
 				foundcharacter = qtrue;
 				ch->skill = token.intvalue;
